+++ contracts/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
 
-from contracts.models import PaymentMethod, Contract
+from contracts.models import Contract
 
-admin.site.register(PaymentMethod)
 admin.site.register(Contract)
