+++ contracts/models.py
@@ -5,36 +5,9 @@ from django.utils.text import slugify
 from django.contrib.auth import get_user_model
 from users.abstracts import UniversalIdModel, TimeStampedModel
 from clients.models import Client
+from payments.models import PaymentMethod
 
 User = get_user_model()
-
-
-class PaymentMethod(UniversalIdModel, TimeStampedModel):
-    user = models.ForeignKey(
-        User, on_delete=models.CASCADE, related_name="payment_methods"
-    )
-    name = models.CharField(max_length=255)
-    description = models.TextField()
-    slug = models.SlugField(max_length=400, unique=True, blank=True, null=True)
-
-    class Meta:
-        verbose_name = "Payment Method"
-        verbose_name_plural = "Payment Methods"
-        ordering = ["-created_at"]
-        constraints = [
-            models.UniqueConstraint(
-                fields=["user", "name"], name="unique_user_payment_method_name"
-            )
-        ]
-
-    def __str__(self):
-        return self.name
-
-
-@receiver(pre_save, sender=PaymentMethod)
-def slug_pre_save(sender, instance, **kwargs) -> None:
-    if instance.slug is None or instance.slug == "":
-        instance.slug = slugify(f"{instance.name}-{instance.id}")
 
 
 class Contract(UniversalIdModel, TimeStampedModel):
@@ -56,6 +29,7 @@ class Contract(UniversalIdModel, TimeStampedModel):
     payment_method = models.ForeignKey(
         PaymentMethod, on_delete=models.CASCADE, related_name="contract_payment"
     )
+    slug = models.SlugField(max_length=400, unique=True, blank=True, null=True)
 
     class Meta:
         verbose_name = "Contract"
@@ -64,3 +38,9 @@ class Contract(UniversalIdModel, TimeStampedModel):
 
     def __str__(self):
         return self.name
+
+
+@receiver(pre_save, sender=Contract)
+def slug_pre_save(sender, instance, **kwargs) -> None:
+    if instance.slug is None or instance.slug == "":
+        instance.slug = slugify(f"{instance.name}-{instance.id}")
