+++ contracts/views.py
@@ -2,31 +2,11 @@ from django.contrib.auth import get_user_model
 from rest_framework import generics
 from rest_framework.permissions import IsAuthenticated
 
-from contracts.serializers import ContractSerializer, PaymentMethodSerializer
-from contracts.models import Contract, PaymentMethod
+from contracts.serializers import ContractSerializer
+from contracts.models import Contract
 
 
 User = get_user_model()
-
-
-class PaymentMethodListCreateView(generics.ListCreateAPIView):
-    serializer_class = PaymentMethodSerializer
-    permission_classes = [IsAuthenticated]
-
-    def get_queryset(self):
-        return PaymentMethod.objects.filter(user=self.request.user)
-
-    def perform_create(self, serializer):
-        serializer.save(user=self.request.user)
-
-
-class PaymentMethodDetailView(generics.RetrieveUpdateDestroyAPIView):
-    serializer_class = PaymentMethodSerializer
-    permission_classes = [IsAuthenticated]
-    lookup_field = "slug"
-
-    def get_queryset(self):
-        return PaymentMethod.objects.filter(user=self.request.user)
 
 
 class ContractListCreateView(generics.ListCreateAPIView):
