+++ payments/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from payments.models import PaymentMethod
+
+
+admin.site.register(PaymentMethod)
