+++ payments/models.py
@@ -1,0 +1,35 @@
+from django.db import models
+from django.db.models.signals import pre_save
+from django.dispatch import receiver
+from django.utils.text import slugify
+from django.contrib.auth import get_user_model
+from users.abstracts import UniversalIdModel, TimeStampedModel
+
+User = get_user_model()
+
+class PaymentMethod(UniversalIdModel, TimeStampedModel):
+    user = models.ForeignKey(
+        User, on_delete=models.CASCADE, related_name="payment_methods"
+    )
+    name = models.CharField(max_length=255)
+    description = models.TextField()
+    slug = models.SlugField(max_length=400, unique=True, blank=True, null=True)
+
+    class Meta:
+        verbose_name = "Payment Method"
+        verbose_name_plural = "Payment Methods"
+        ordering = ["-created_at"]
+        constraints = [
+            models.UniqueConstraint(
+                fields=["user", "name"], name="unique_user_payment_method_name"
+            )
+        ]
+
+    def __str__(self):
+        return self.name
+
+
+@receiver(pre_save, sender=PaymentMethod)
+def slug_pre_save(sender, instance, **kwargs) -> None:
+    if instance.slug is None or instance.slug == "":
+        instance.slug = slugify(f"{instance.name}-{instance.id}")
