+++ suite/settings/base.py
@@ -53,6 +53,7 @@ INSTALLED_APPS = [
     "users",
     "clients",
     "contracts",
+    "payments",
 ]
 
 MIDDLEWARE = [
